Added a few more events
